backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-multipart
jinja2
weasyprint
google-generativeai
python-dotenv
